+++ api/src/environment/dto/update-environment.dto.ts
@@ -1,8 +1,10 @@
 import { PartialType } from '@nestjs/swagger';
+import { DagDto } from 'src/reconciliation/dtos/environment-dag.dto';
 import { CreateEnvironmentDto } from './create-environment.dto';
 
 export class UpdateEnvironmentDto extends PartialType(CreateEnvironmentDto) {
   name: string;
-  teamName: string;
   duration: number;
+  isDeleted: boolean;
+  dag: DagDto;
 }