+++ api/src/environment/environment.controller.ts
@@ -1,14 +1,42 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Request } from '@nestjs/common';
 import { EnvironmentService } from './environment.service';
-import { CreateEnvironmentDto } from './dto/create-environment.dto';
 import { UpdateEnvironmentDto } from './dto/update-environment.dto';
+import { ComponentService } from 'src/costing/services/component.service';
 
-@Controller('environment')
+@Controller({
+  version: '1'
+})
 export class EnvironmentController {
-  constructor(private readonly envSvc: EnvironmentService) {}
+  constructor(
+    private readonly envSvc: EnvironmentService,
+    private readonly compSvc: ComponentService
+    ) {}
 
-  @Post()
-  create(@Body() createEnvironmentDto: CreateEnvironmentDto) {
-    return this.envSvc.create(createEnvironmentDto);
+  @Get()
+  async findOne(@Request() req) {
+    const {org, team, env} = req;
+
+    return this.envSvc.findById(org, env.id, team);
+  }
+
+  @Get('/cost')
+  async getTeamCost(@Request() req) {
+    const {org, team, env} = req;
+    
+    return this.compSvc.getEnvironmentCost(org, team, env.name);
+  }
+
+  @Patch()
+  async update(@Request() req, @Param('id') id: number, @Body() updateEnvDto: UpdateEnvironmentDto) {
+    const {org, team, env} = req;
+
+    return this.envSvc.update(org, id, updateEnvDto);
+  }
+
+  @Delete()
+  remove(@Request() req, @Param('id') id: number) {
+    const {org, team, env} = req;
+
+    return this.envSvc.remove(org, id);
   }
 }