+++ api/src/reconciliation/reconciliation.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ComponentService } from 'src/costing/services/component.service';
 import { EnvironmentService } from 'src/environment/environment.service';
+import { RootEnvironmentService } from 'src/root-environment/root.environment.service';
 import { TeamService } from 'src/team/team.service';
 import { Component, ComponentReconcile, Environment, EnvironmentReconcile, Team } from 'src/typeorm';
 import { ReconciliationController } from './reconciliation.controller';
@@ -24,6 +25,7 @@ import { SSEService } from './sse.service';
   ],
   providers: [
     ReconciliationService,
+    RootEnvironmentService,
     EnvironmentService,
     SSEService,
     TeamService,