+++ api/src/root-team/root.team.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
-import { RootTeamService } from './root-team.service';
-import { RootTeamController } from './root-team.controller';
+import { RootTeamService } from './root.team.service';
+import { RootTeamController } from './root.team.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Team } from 'src/typeorm';
+import { TeamService } from 'src/team/team.service';
 
 @Module({
   imports: [
@@ -12,7 +13,8 @@ import { Team } from 'src/typeorm';
   ],
   controllers: [RootTeamController],
   providers: [
-    RootTeamService
+    RootTeamService,
+    TeamService
   ]
 })
 export class RootTeamModule {}