+++ api/src/routes.ts
@@ -2,14 +2,16 @@ import { Routes } from "@nestjs/core";
 import { AuthModule } from "./auth/auth.module";
 import { CostingModule } from "./costing/costing.module";
 import { OrganizationModule } from "./organization/organization.module";
-import { RootOrganizationsModule } from "./rootOrganization/rootOrganization.module";
+import { RootOrganizationsModule } from "./root-organization/root.organization.module";
 import { ReconciliationModule } from "./reconciliation/reconciliation.module";
 import { SecretsModule } from "./secrets/secrets.module";
 import { UsersModule } from "./users/users.module";
 import { SystemModule } from "./system/system.module";
 import { OperationsModule } from "./operations/operations.module";
 import { TeamModule } from "./team/team.module";
-import { RootTeamModule } from "./root-team/root-team.module";
+import { RootTeamModule } from "./root-team/root.team.module";
+import { EnvironmentModule } from "./environment/environment.module";
+import { RootEnvironmentModule } from "./root-environment/root.environment.module";
 
 export const appRoutes: Routes = [
   {
@@ -49,12 +51,22 @@ export const appRoutes: Routes = [
                 module: TeamModule,
                 children: [
                   {
-                    path: 'costing',
-                    module: CostingModule
-                  },
-                  {
-                    path: 'reconciliation',
-                    module: ReconciliationModule
+                    path: 'environments',
+                    module: RootEnvironmentModule,
+                    children: [
+                      {
+                        path: '/:environmentId',
+                        module: EnvironmentModule
+                      },
+                      {
+                        path: 'costing',
+                        module: CostingModule
+                      },
+                      {
+                        path: 'reconciliation',
+                        module: ReconciliationModule
+                      },
+                    ]
                   },
                 ]
               }