+++ api/src/team/dto/team-spec.dto.ts
@@ -1,0 +1,9 @@
+export class TeamSpecDto {
+  teamName: string
+  configRepo: TeamConfigRepoDto
+}
+
+export class TeamConfigRepoDto {
+  source: string
+  path: string
+}