+++ api/src/types.ts
@@ -1,8 +1,8 @@
 import { Request } from "express";
-import { Team } from "./typeorm";
+import { Environment, Team } from "./typeorm";
 import { Organization } from "./typeorm/Organization.entity";
 
-export type APIRequest = Request & { org: Organization, team: Team }
+export type APIRequest = Request & { org: Organization, team: Team, env?: Environment }
 
 export const SqlErrorCodes = {
   NO_DEFAULT: 'ER_NO_DEFAULT_FOR_FIELD',