+++ api/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Created
 
-* added `orgs/:orgId/teams/:teamId` routing
+* teams `orgs/:orgId/teams/:teamId` routing
+* environments `orgs/:orgId/teams/:teamId/environments/:environmentId` routing
 
 ### Modified
 
@@ -17,7 +18,8 @@
 * `POST reconciliation/component/save` - all components are now saved via `POST /orgs/:orgId/team/:teamId/reconciliation/spec`
 * `POST reconciliation/environment/save` - all environments are now saved via `POST /orgs/:orgId/team/:teamId/reconciliation/spec`
 * `POST costing/savecomponent` - all components are now saved via `POST /orgs/:orgId/team/:teamId/reconciliation/spec`
-* `GET costing/team/:name` - use `GET teams/:teamId/cost`
+* `GET costing/team/:name` - use `GET teams/:teamId?withCost=true`
+* `GET costing/environment/:name` - use `GET teams/:teamId/environments/:environmentId?withCost=true`
 
 ## Multi-tenancy Changes